+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r200-release...35b0ccfd96956e313858d81e4d3da795e963fa34)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r201-release...85f14f4c0b97092bf9fb69eb8a32813f5e8ac343)
 
-- Update Amirdrassil debuffs
-- Update Health Text
+- Add support for ping system (retail)
+- Expand Spotlights
+- Add "Show stack text" option
+- Add a shortcut to refresh all indicators
+- Fix ResetCustomIndicators
+- Fix Buff Tracker (retail)
+- Fix UpdatePixelPerfect event
+- Try to fix a Cell initialization issue
+- Update OmniCD support
+- [snippet] SwingTimer